requirements.txt
# Core dependencies
python-telegram-bot==20.7
python-dotenv==1.0.0
pydantic==2.5.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
pandas==2.1.4

# LLM & AI
openai==1.6.1
anthropic==0.7.8
google-generativeai==0.3.2
langchain==0.1.0
langchain-openai==0.0.2

# RAG & Embeddings
chromadb==0.4.22
sentence-transformers==2.2.2
faiss-cpu==1.7.4

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0
Pillow==10.1.0

# Utilities
pyyaml==6.0.1
structlog==23.3.0
prometheus-client==0.19.0
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
faker==22.0.0

# Development
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0

# Optional: Local LLM support
# transformers==4.36.2
# torch==2.1.2
# accelerate==0.25.0